requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.1.2
django-ckeditor==5.6.1
django-classy-tags==0.8.0
django-js-asset==1.1.0
django-modeltranslation==0.13b1
django-recaptcha2==1.3.0
django-rosetta==0.9.0
django-suit==0.2.26
django-taggit==0.23.0
django-taggit-templatetags2==1.6.1
gunicorn==19.9.0
idna==2.7
Markdown==3.0.1
Pillow==5.2.0
polib==1.1.0
psycopg2-binary==2.7.5
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
